refactor: LottoNumbers를 LottoTicket에 포함
