+++ src/main/java/lotto/domain/lotto/strategy/LottoNumbersCreateStrategy.java
@@ -1,8 +1,10 @@
 package lotto.domain.lotto.strategy;
 
-import lotto.domain.lotto.LottoNumbers;
+import lotto.domain.lotto.LottoNumber;
+
+import java.util.List;
 
 @FunctionalInterface
 public interface LottoNumbersCreateStrategy {
-    LottoNumbers create();
+    List<LottoNumber> create();
 }
