+++ src/main/java/lotto/domain/lotto/strategy/RandomNumbersCreateStrategy.java
@@ -1,27 +1,28 @@
 package lotto.domain.lotto.strategy;
 
 import lotto.domain.lotto.LottoNumber;
-import lotto.domain.lotto.LottoNumbers;
 import lotto.domain.lotto.LottoTicket;
 
 import java.util.ArrayList;
 import java.util.List;
 
+import static lotto.domain.lotto.LottoTicket.LOTTO_NUMBERS_LENGTH;
+
 public class RandomNumbersCreateStrategy implements LottoNumbersCreateStrategy {
     private static final RandomNumberCreateStrategy createRandomNumberStrategy = new RandomNumberCreateStrategy();
 
     @Override
-    public LottoNumbers create() {
+    public List<LottoNumber> create() {
         List<LottoNumber> lottoNumbers = new ArrayList<>();
-        while (lottoNumbers.isEmpty() || LottoNumbers.hasDuplicates(lottoNumbers)) {
+        while (lottoNumbers.isEmpty() || LottoTicket.hasDuplicates(lottoNumbers)) {
             lottoNumbers = createRandomNumbers();
         }
-        return LottoNumbers.of(lottoNumbers);
+        return lottoNumbers;
     }
 
     private List<LottoNumber> createRandomNumbers() {
         final List<LottoNumber> lottoNumbers = new ArrayList<>();
-        for (int i = 0; i < LottoNumbers.LOTTO_NUMBERS_LENGTH; i++) {
+        for (int i = 0; i < LOTTO_NUMBERS_LENGTH; i++) {
             lottoNumbers.add(createRandomNumberStrategy.create());
         }
         return lottoNumbers;
