+++ src/test/java/lotto/domain/lotto/strategy/RandomNumbersCreateStrategyTest.java
@@ -1,8 +1,10 @@
 package lotto.domain.lotto.strategy;
 
 import lotto.domain.lotto.LottoNumber;
-import lotto.domain.lotto.LottoNumbers;
+import lotto.domain.lotto.LottoTicket;
 import org.junit.jupiter.api.Test;
+
+import java.util.List;
 
 import static org.assertj.core.api.AssertionsForInterfaceTypes.assertThat;
 
@@ -11,10 +13,10 @@ class RandomNumbersCreateStrategyTest {
     void create() {
         // given
         final LottoNumbersCreateStrategy strategy = new RandomNumbersCreateStrategy();
-        final int lottoNumberLength = LottoNumbers.LOTTO_NUMBERS_LENGTH;
+        final int lottoNumberLength = LottoTicket.LOTTO_NUMBERS_LENGTH;
 
         // when
-        final LottoNumbers lottoNumbers = strategy.create();
+        final List<LottoNumber> lottoNumbers = strategy.create();
 
         // then
         assertThat(lottoNumbers).isNotNull();
